rename main folder and update readme
