+++ nautilus_nlp/__init__.py
@@ -15,4 +15,4 @@
 # You should have received a copy of the GNU Lesser General Public License
 # along with this program; if not, write to the Free Software Foundation,
 # Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
-from preprocessing.preprocessor import Preprocessor
+from nautilus_nlp.preprocessor import Preprocessor
