+++ nautilus_nlp/preprocessor.py
@@ -3,9 +3,9 @@ from typing import List, Callable
 from sklearn.pipeline import Pipeline
 from sklearn.preprocessing import FunctionTransformer
 
-from preprocessing.social.preprocess import (
+from nautilus_nlp.social.preprocess import (
     remove_html_tags, remove_mentions, remove_emoji, remove_hashtag)
-from preprocessing.classic.preprocess import normalize_whitespace, remove_eol_characters, fix_bad_unicode
+from nautilus_nlp.classic.preprocess import normalize_whitespace, remove_eol_characters, fix_bad_unicode
 
 
 class Preprocessor():
