+++ nautilus_nlp/social/preprocess.py
@@ -21,7 +21,7 @@ from __future__ import absolute_import, division, print_function, unicode_litera
 
 import emoji as _emoji
 from config import constants
-from preprocessing.classic.preprocess import normalize_whitespace
+from nautilus_nlp.classic.preprocess import normalize_whitespace
 
 
 def remove_mentions(text) -> str:
